Add files via upload
As a Data Scientist for Hass Consulting Company which is a real estate leader with over 25 years of experience. The task is to study the factors that affect housing prices using the given information on real estate properties that was collected over the past few months. Later onwards, create a model that would allow the company to accurately predict the sale of prices upon being provided with the predictor variables.